requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.4
colorlog==6.6.0
english-words==1.0.4
frozenlist==1.3.0
hikari==2.0.0.dev108
hikari-lightbulb==2.2.2
idna==3.3
install==1.3.5
multidict==6.0.2
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
windows-curses==2.3.0
yarl==1.7.2
